pyproject.toml
[tool.poetry]
name = "ocr-web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"
requests = "^2.32.3"
python-multipart = "^0.0.17"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
